Refactored attachments component.
