+++ src/components/flow/actions/sendmsg/SendMsgForm.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/explicit-member-accessibility */
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import { react as bindCallbacks } from 'auto-bind';
-import axios, { AxiosResponse } from 'axios';
+import { AxiosResponse } from 'axios';
 import Dialog, { ButtonSet, Tab } from 'components/dialog/Dialog';
 import { hasErrors, renderIssues } from 'components/flow/actions/helpers';
 import {
@@ -20,7 +20,7 @@ import TextInputElement from 'components/form/textinput/TextInputElement';
 import HelpIcon from 'components/helpicon/HelpIcon';
 import TypeList from 'components/nodeeditor/TypeList';
 import { fakePropType } from 'config/ConfigProvider';
-import { fetchAsset, getCookie } from 'external';
+import { fetchAsset } from 'external';
 import { Template, TemplateTranslation } from 'flowTypes';
 import mutate from 'immutability-helper';
 import * as React from 'react';
@@ -34,20 +34,17 @@ import {
   FormEntry
 } from 'store/nodeEditor';
 import { MaxOfTenItems, Required, shouldRequireIf, validate } from 'store/validators';
-import { createUUID, range, renderIf } from 'utils';
+import { range, renderIf } from 'utils';
 
 import styles from './SendMsgForm.module.scss';
 import { hasFeature } from 'config/typeConfigs';
 import { FeatureFilter, FlowTypes } from 'config/interfaces';
-import Pill from 'components/pill/Pill';
 
 import variables from 'variables.module.scss';
 
 import i18n from 'config/i18n';
 import { Trans } from 'react-i18next';
 import { Attachment, renderAttachments } from './attachments';
-import Select from 'react-select/base';
-import { small } from '../../../../utils/reactselect';
 
 const FACEBOOK_ICON = require('static/images/facebook.png');
 const TELEGRAM_ICON = require('static/images/telegram.png');
@@ -84,8 +81,6 @@ export interface SendMsgFormState extends FormState {
 }
 
 export default class SendMsgForm extends React.Component<ActionFormProps, SendMsgFormState> {
-  private filePicker: any;
-
   constructor(props: ActionFormProps, context: any) {
     super(props);
     this.context = context;
@@ -254,114 +249,6 @@ export default class SendMsgForm extends React.Component<ActionFormProps, SendMs
     };
   }
 
-  private renderUpload(index: number, attachment: Attachment): JSX.Element {
-    return (
-      <div
-        className={styles.url_attachment}
-        key={index > -1 ? 'url_attachment_' + index : createUUID()}
-      >
-        <div className={styles.type_choice}>
-          <Select
-            name="Type"
-            styles={small as any}
-            className={styles.selected_file}
-            value={{ label: attachment.type }}
-            isDisabled={true}
-          />
-        </div>
-        <div className={styles.url}>
-          <span className={styles.upload}>
-            <Pill
-              icon="fe-download"
-              text="Download"
-              large={true}
-              onClick={() => {
-                window.open(attachment.url, '_blank');
-              }}
-            />
-            <div className={styles.remove_upload}>
-              <Pill
-                icon="fe-x"
-                text="Remove"
-                large={true}
-                onClick={() => {
-                  this.handleAttachmentRemoved(index);
-                }}
-              />
-            </div>
-          </span>
-        </div>
-      </div>
-    );
-  }
-
-  private handleUploadFile(files: FileList): void {
-    let attachments: any = this.state.attachments;
-
-    // if we have a csrf in our cookie, pass it along as a header
-    const csrf = getCookie('csrftoken');
-    const headers: any = csrf ? { 'X-CSRFToken': csrf } : {};
-
-    // mark us as ajax
-    headers['X-Requested-With'] = 'XMLHttpRequest';
-
-    if (!this.isAttachmentsValid(files)) {
-      return null;
-    }
-
-    const data = new FormData();
-    data.append('file', files[0]);
-    axios
-      .post(this.context.config.endpoints.attachments, data, { headers })
-      .then(response => {
-        attachments = mutate(attachments, {
-          $push: [{ type: response.data.type, url: response.data.url, uploaded: true }]
-        });
-        this.setState({ attachments });
-      })
-      .catch(error => {
-        console.log(error);
-      });
-  }
-
-  private isAttachmentsValid(files: FileList) {
-    let title = '';
-    let message = '';
-    let isValid = true;
-    const file = files[0];
-    const fileType = file.type.split('/')[0];
-    const fileEncoding = file.name.split('.').pop();
-
-    if (file.type !== 'application/pdf' && !['audio', 'video', 'image'].includes(fileType)) {
-      title = 'Invalid Attachment';
-      message = 'Attachments must be either PDF, video, audio, or an image.';
-      isValid = false;
-    } else if (
-      fileType === 'audio' &&
-      !['mp3', 'm4a', 'x-m4a', 'wav', 'ogg', 'oga'].includes(fileEncoding)
-    ) {
-      title = 'Invalid Format';
-      message = 'Audio attachments must be encoded as mp3, m4a, wav, ogg or oga files.';
-      isValid = false;
-    } else if ((fileType === 'image' && file.size > 512000) || file.size > 20971520) {
-      title = 'File Size Exceeded';
-      message =
-        'The file size should be less than 500kB for images and less than 20MB for audio, video and PDF files. Please choose another file and try again.';
-      isValid = false;
-    }
-
-    if (!isValid) {
-      this.props.mergeEditorState({
-        modalMessage: {
-          title: title,
-          body: message
-        },
-        saving: false
-      });
-    }
-    return isValid;
-  }
-
   private handleTemplateChanged(selected: any[]): void {
     const template = selected ? selected[0] : null;
 
@@ -506,6 +393,16 @@ export default class SendMsgForm extends React.Component<ActionFormProps, SendMs
       $splice: [[index, 1]]
     });
     this.setState({ attachments });
+  }
+
+  private handleAttachmentError(title: string, message: string) {
+    this.props.mergeEditorState({
+      modalMessage: {
+        title: title,
+        body: message
+      },
+      saving: false
+    });
   }
 
   private handleAddQuickReply(newQuickReply: string): boolean {
@@ -804,7 +701,8 @@ export default class SendMsgForm extends React.Component<ActionFormProps, SendMs
         this.state.attachments,
         this.handleAttachmentUploaded,
         this.handleAttachmentChanged,
-        this.handleAttachmentRemoved
+        this.handleAttachmentRemoved,
+        this.handleAttachmentError
       ),
       checked: this.state.attachments.length > 0
     };
